Renames middleware to transformers and textMW to textResponse
