+++ packages/unmock-core/src/__tests__/middleware.test.ts
@@ -1,4 +1,4 @@
-import { middleware } from "../index";
+import { transformers } from "../index";
 import { Schema } from "../service/interfaces";
 
 const schema: Schema = {
@@ -34,7 +34,7 @@ const schema: Schema = {
 
 describe("Test text provider", () => {
   it("returns empty object for undefined state", () => {
-    const p = middleware.textMW();
+    const p = transformers.textResponse();
     expect(p.isEmpty).toBeTruthy();
     expect(p.top).toEqual({});
     // @ts-ignore // deliberately checking with empty input
@@ -42,7 +42,7 @@ describe("Test text provider", () => {
   });
 
   it("returns empty object for empty state", () => {
-    const p = middleware.textMW("");
+    const p = transformers.textResponse("");
     expect(p.isEmpty).toBeTruthy();
     expect(p.top).toEqual({});
     // @ts-ignore // deliberately checking with empty input
@@ -50,7 +50,7 @@ describe("Test text provider", () => {
   });
 
   it("returns empty object for empty schema", () => {
-    const p = middleware.textMW("foo");
+    const p = transformers.textResponse("foo");
     expect(p.isEmpty).toBeFalsy();
     expect(p.top).toEqual({});
     // @ts-ignore // deliberately checking with empty input
@@ -58,14 +58,14 @@ describe("Test text provider", () => {
   });
 
   it("returns empty object for non-text schema", () => {
-    const p = middleware.textMW("foo");
+    const p = transformers.textResponse("foo");
     expect(p.isEmpty).toBeFalsy();
     expect(p.top).toEqual({});
     expect(p.gen({ type: "array", items: {} })).toEqual({});
   });
 
   it("returns correct state object for valid input", () => {
-    const p = middleware.textMW("foo");
+    const p = transformers.textResponse("foo");
     expect(p.isEmpty).toBeFalsy();
     expect(p.top).toEqual({});
     expect(p.gen({ type: "string" })).toEqual({ type: "string", const: "foo" });
@@ -73,7 +73,7 @@ describe("Test text provider", () => {
 
   it("top level DSL doesn't change response", () => {
     // @ts-ignore // invalid value on purpose
-    const p = middleware.textMW("foo", { $code: 200, notDSL: "a" });
+    const p = transformers.textResponse("foo", { $code: 200, notDSL: "a" });
     expect(p.isEmpty).toBeFalsy();
     expect(p.top).toEqual({ $code: 200 }); // non DSL is filtered out
     expect(p.gen({ type: "string" })).toEqual({ type: "string", const: "foo" });
@@ -82,21 +82,21 @@ describe("Test text provider", () => {
 
 describe("Test default provider", () => {
   it("returns empty objects for undefined state", () => {
-    const p = middleware.default();
+    const p = transformers.default();
     expect(p.isEmpty).toBeTruthy();
     expect(p.top).toEqual({});
     expect(p.gen({})).toEqual({});
   });
 
   it("returns empty objects for empty state", () => {
-    const p = middleware.default({});
+    const p = transformers.default({});
     expect(p.isEmpty).toBeTruthy();
     expect(p.top).toEqual({});
     expect(p.gen({})).toEqual({});
   });
 
   it("filters out top level DSL from state", () => {
-    const p = middleware.default({ $code: 200, foo: "bar" });
+    const p = transformers.default({ $code: 200, foo: "bar" });
     expect(p.isEmpty).toBeFalsy();
     expect(p.top).toEqual({ $code: 200 });
     expect(p.gen({})).toEqual({}); // no schema to expand
@@ -111,12 +111,12 @@ describe("Test default provider", () => {
   });
 
   it("with empty path", () => {
-    const spreadState = middleware.default({}).gen(schema);
+    const spreadState = transformers.default({}).gen(schema);
     expect(spreadState).toEqual({}); // Empty state => empty spread state
   });
 
   it("with specific path", () => {
-    const spreadState = middleware
+    const spreadState = transformers
       .default({
         test: { id: "a" },
       })
@@ -134,13 +134,13 @@ describe("Test default provider", () => {
   });
 
   it("with vague path", () => {
-    const spreadState = middleware.default({ id: 5 }).gen(schema);
+    const spreadState = transformers.default({ id: 5 }).gen(schema);
     // no "id" in top-most level or immediately under properties\items
     expect(spreadState).toEqual({ id: null });
   });
 
   it("with missing parameters", () => {
-    const spreadState = middleware.default({ ida: "a" }).gen(schema);
+    const spreadState = transformers.default({ ida: "a" }).gen(schema);
     expect(spreadState).toEqual({ ida: null }); // Nothing to spread
   });
 });
