+++ packages/unmock-node/src/__tests__/backend/states.test.ts
@@ -1,7 +1,6 @@
 import axios from "axios";
 import path from "path";
-import { UnmockOptions } from "unmock-core";
-import { textMW } from "unmock-core/dist/service/state/middleware";
+import { transformers, UnmockOptions } from "unmock-core";
 import NodeBackend from "../../backend";
 
 const servicesDirectory = path.join(__dirname, "..", "loaders", "resources");
@@ -71,16 +70,16 @@ describe("Node.js interceptor", () => {
     });
 
     test("gets correct state when setting textual middleware", async () => {
-      states.petstore(textMW("foo"));
+      states.petstore(transformers.textResponse("foo"));
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.status).toBe(200);
       expect(response.data).toBe("foo");
     });
 
     test("throws when setting textual middleware with DSL with non-existing status code", async () => {
-      expect(() => states.petstore(textMW("foo", { $code: 400 }))).toThrow(
-        "status code '400'",
-      );
+      expect(() =>
+        states.petstore(transformers.textResponse("foo", { $code: 400 })),
+      ).toThrow("status code '400'");
     });
   });
 });
